+++ tests/test_group_no_pain_no_gain/links.py
@@ -2,3 +2,4 @@ HOME_PAGE = 'https://openweathermap.org/'
 PARTNERS_AND_SOLUTIONS = 'https://openweathermap.org/examples'
 COOKIES_SETTINGS = 'https://openweathermap.org/cookies-settings'
 GUIDE_PAGE = 'https://openweathermap.org/guide'
+ASK_A_QUESTION_PAGE = "https://home.openweathermap.org/questions"
